Made ErDatabase app with fab button that pops open a modal and blurs background
